Revert "Move QrCodeGenerator to logic"
This reverts commit 84e1876a129236b16d97b365a7b5961e4a7f6d84.
