+++ src/main/java/seedu/address/model/person/Person.java
@@ -12,9 +12,9 @@ import java.util.logging.Logger;
 
 import com.google.zxing.WriterException;
 
+import seedu.address.QrCodeGenerator;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.logic.QrCodeGenerator;
 import seedu.address.model.person.exceptions.AttributeNotFoundException;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.TagSet;
@@ -25,15 +25,30 @@ import seedu.address.model.tag.TagSet;
  * immutable.
  */
 public class Person {
+    /**
+     * Attributes of a Person
+     */
+    public enum PersonAttribute {
+        NAME,
+        PHONE,
+        EMAIL,
+        ADDRESS,
+        TAGS,
+        NOTE
+    }
+
     private static final Logger logger = LogsCenter.getLogger(Person.class);
+
     // Identity fields
     private final Name name;
     private final Phone phone;
     private final Email email;
+
     // Data fields
     private final Address address;
     private final TagSet tags;
     private final Note note;
+
     /**
      * Every field must be present and not null.
      */
@@ -53,6 +68,7 @@ public class Person {
      * Get the valued of the specified attribute.
      *
      * @param attribute Attribute to retrieve
+     *
      * @return Value of the specified attribute
      */
     public Attribute<? extends Object> getAttribute(PersonAttribute attribute) {
@@ -198,17 +214,5 @@ public class Person {
             logger.warning("Unable to delete QR code for " + this);
             return false;
         }
-    }
-
-    /**
-     * Attributes of a Person
-     */
-    public enum PersonAttribute {
-        NAME,
-        PHONE,
-        EMAIL,
-        ADDRESS,
-        TAGS,
-        NOTE
     }
 }
