+++ src/test/java/seedu/address/QrCodeGeneratorTest.java
@@ -12,8 +12,6 @@ import org.junit.jupiter.api.Test;
 
 import com.google.zxing.WriterException;
 
-import seedu.address.logic.QrCodeGenerator;
-
 class QrCodeGeneratorTest {
 
     @Test
